+++ README.md
@@ -27,6 +27,7 @@
   - [x] 플레이어의 턴이 끝이 끝나면 딜러의 턴이 시작된다.
     - [x] 딜러의 카드 합계가 16이하이면 카드를 더 받는다.
   - [x] 플레이어와 딜러의 턴이 끝나면 Result를 반환한다.
+- [x] 플레이어와 딜러의 턴을 관리하는 TurnMachine 추가
 - [ ] 승패의 결과는 최종에서 결정된다.
   - [x] 딜러는 17미만이면 무조건 카드를 더 받는다.
 - [x] Participant을 리팩터링한다.
