+++ src/main/kotlin/blackjack/BlackJackGame.kt
@@ -10,32 +10,13 @@ import blackjack.view.OutputView
 data class BlackJackGame(
     val dealer: Dealer,
     val gamePlayers: Players,
-    val deck: Deck,
-    val moreCardPolicy: InputMoreCardPolicy,
+    val turnMachine: TurnMachine,
 ) {
     fun play(): List<ParticipantResult> {
-        gamePlayers.players.forEach { player -> playerTurn(player) }
-        dealerTurn()
+        gamePlayers.players.forEach { player -> turnMachine.playerTurn(player) }
+        turnMachine.dealerTurn(dealer)
 
         return ParticipantResult.of(dealer, gamePlayers)
-    }
-
-    private fun playerTurn(player: Player) {
-        while (!player.isBust()) {
-            val isMoreCard = moreCardPolicy.isMoreCard(player.getName())
-            if (!isMoreCard) {
-                break
-            }
-            player.take(listOf(deck.pick()))
-            OutputView.printPlayerCard(player)
-        }
-    }
-
-    private fun dealerTurn() {
-        if(dealer.isMoreCard()) {
-            dealer.take(listOf(deck.pick()))
-            OutputView.printDealerMoreCard()
-        }
     }
 
     companion object {
@@ -43,12 +24,15 @@ data class BlackJackGame(
             dealer: Dealer,
             players: List<Player>,
             deck: Deck,
-            moreCardPolicy: InputMoreCardPolicy,
         ): BlackJackGame {
             dealer.take(listOf(deck.pick(), deck.pick()))
             players.forEach { player -> player.take(listOf(deck.pick(), deck.pick())) }
 
-            return BlackJackGame(dealer, Players(players), deck, moreCardPolicy)
+            return BlackJackGame(
+                dealer,
+                Players(players),
+                TurnMachine(deck, InputMoreCardPolicy, OutputView),
+            )
         }
     }
 }
